@@ -1,7 +1,7 @@
 source 'https://rubygems.org'
 ruby '2.6.3'
 
-gem 'rails', '~> 5.2'
+gem 'rails', '6.0.0'
 gem 'bootstrap'
 gem 'attr_encrypted'
 gem 'jquery-rails'
@@ -57,7 +57,6 @@ group :test do
   gem 'webmock'
   gem 'mocha'
   gem 'minitest'
-  gem 'action-cable-testing'
   gem 'timecop'
 end
 
