@@ -6,6 +6,6 @@ class SyncGithubAppAuthorizationWorker
 
   def perform(github_id)
     user = User.find_by_github_id(github_id)
-    user.update_attributes(app_token: nil) if user.present?
+    user.update(app_token: nil) if user.present?
   end
 end
