@@ -49,9 +49,10 @@ module Octobox
         current_subs_purchases.each do |purchase|
 
           next if purchase.app_installation.nil?
+
           next if subscriber_names.include? purchase.app_installation.account_login
           unless purchase.next_billing_date.present? && purchase.next_billing_date > Time.now
-            purchase.update_attributes(unit_count: 0)
+            purchase.update(unit_count: 0)
             Rails.logger.info("n\n\033[32m[#{Time.current}] INFO -- Removed #{plan_name} for #{purchase.app_installation.account_login}\033[0m\n\n")
           end
         end
@@ -68,10 +69,10 @@ module Octobox
           app_installation.create_subscription_purchase(subscription_plan: plan, unit_count: 1)
           Rails.logger.info("n\n\033[32m[#{Time.current}] INFO -- Added #{plan_name} for #{subscriber}\033[0m\n\n")
         elsif subscription_purchase.subscription_plan.name == plan_name && subscription_purchase.unit_count.zero?
-          subscription_purchase.update_attributes(unit_count: 1)
+          subscription_purchase.update(unit_count: 1)
           Rails.logger.info("n\n\033[32m[#{Time.current}] INFO -- Restarted #{plan_name} for #{subscriber}\033[0m\n\n")
         else
-          subscription_purchase.update_attributes(subscription_plan: plan, unit_count: 1)
+          subscription_purchase.update(subscription_plan: plan, unit_count: 1)
           Rails.logger.info("n\n\033[32m[#{Time.current}] INFO -- Switched #{subscriber} to #{plan_name}\033[0m\n\n")
         end
       end
